pyproject.toml
[tool.poetry]
name = "gbmargindashboard"
version = "0.1.0"
description = ""
authors = ["Imran <imran@personal>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.13"
pandas = "^2.1.1"
plotly = "^5.18.0"
streamlit = "^1.27.2"
requests = "^2.31.0"
entsoe-py = "^0.5.10"
stqdm = "^0.0.5"
xmltodict = "^0.13.0"
jupyter = "^1.0.0"
notebook = "^7.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
